added paste but do not merge until unit testing
